Improve monthly and annual metadata vignette
Combined id/column name and description and added search feature.

This was the raw files and completely unstructured, I decided to break this down by measures and breakdowns, which is a much clearer structure in my opinion.

I also decided to switch from gt to reactable, as it as far more customisation options, I think this will be very useful for presenting all the metadata in a clearer way.